MailMethodsTraitおよびTemplateMethodsTraitを追加し、cfFormMailerクラスのメール送信機能を強化
