+++ includes/class.cfFormMailer.inc.php
@@ -13,11 +13,15 @@
 // トレイトの読み込み
 require_once __DIR__ . '/traits/ValidationMethodsTrait.php';
 require_once __DIR__ . '/traits/FilterMethodsTrait.php';
+require_once __DIR__ . '/traits/TemplateMethodsTrait.php';
+require_once __DIR__ . '/traits/MailMethodsTrait.php';
 
 class Class_cfFormMailer
 {
     use ValidationMethodsTrait;
     use FilterMethodsTrait;
+    use TemplateMethodsTrait;
+    use MailMethodsTrait;
 
     public $cfg = array();
 
@@ -620,240 +624,6 @@ class Class_cfFormMailer
         return $rs;
     }
 
-    private function sendAdminMail()
-    {
-        $this->debugLog('管理者宛メール送信処理を開始');
-
-        $reply_to = $this->getAutoReplyAddress();
-
-        // 管理者宛メールの本文生成
-        $tmpl = $this->loadTemplate($this->config('tmpl_mail_admin'));
-        if (!$tmpl) {
-            $this->setError('メールテンプレートの読み込みに失敗しました');
-            return false;
-        }
-
-        $admin_addresses = $this->makeAdminAddress();
-        $admin_address = $admin_addresses[0];
-
-        // 管理者宛送信
-        evo()->loadExtension('MODxMailer');
-        $pm = evo()->mail;
-        foreach ($admin_addresses as $v) {
-            $pm->AddAddress($v);
-        }
-        if ($this->config('admin_mail_cc')) {
-            foreach (explode(',', $this->config('admin_mail_cc')) as $v) {
-                $v = trim($v);
-                if ($this->_isValidEmail($v)) {
-                    $pm->AddCC($v);
-                }
-            }
-        }
-        if ($this->config('admin_mail_bcc')) {
-            foreach (explode(',', $this->config('admin_mail_bcc')) as $v) {
-                $v = trim($v);
-                if ($this->_isValidEmail($v)) {
-                    $pm->AddBCC($v);
-                }
-            }
-        }
-
-        $pm->Subject = $this->config('admin_subject')
-            ? evo()->parseText($this->config('admin_subject'), $this->form)
-            : 'サイトから送信されたメール';
-
-        $pm->setFrom(
-            $admin_address,
-            $this->config('admin_name')
-                ? evo()->parseText($this->config('admin_name'), $this->form)
-                : ''
-        );
-        if ($reply_to) {
-            $pm->addReplyTo($reply_to);
-        }
-        $pm->Sender = $pm->From;
-        $pm->Body = $this->makeBody(
-            $tmpl,
-            ($this->makePh($this->form, $admin_address))
-        );
-        $pm->Encoding = '7bit';
-
-        // ユーザーからのファイル送信
-        $upload_flag = false;
-        $uploaded = sessionv('_cf_uploaded');
-        if (is_array($uploaded) && count($uploaded)) {
-            foreach ($uploaded as $attach_file) {
-                if (!is_file($attach_file['path'])) {
-                    continue;
-                }
-                $pm->AddAttachment(
-                    $attach_file['path'],
-                    mb_convert_encoding(
-                        urldecode(basename($attach_file['path'])),
-                        $this->config('mail_charset'),
-                        $this->config('charset')
-                    )
-                );
-                if (!$upload_flag) {
-                    $upload_flag = true;
-                }
-            }
-        }
-
-        // デバッグログ: メール送信情報
-        if ($this->config('debug_mode')) {
-            $debug_info = "管理者宛メール送信を試行\n";
-            $debug_info .= "宛先: " . implode(', ', $admin_addresses) . "\n";
-            if ($this->config('admin_mail_cc')) {
-                $debug_info .= "CC: " . $this->config('admin_mail_cc') . "\n";
-            }
-            if ($this->config('admin_mail_bcc')) {
-                $debug_info .= "BCC: " . $this->config('admin_mail_bcc') . "\n";
-            }
-            $debug_info .= "件名: " . $pm->Subject . "\n";
-            $debug_info .= "送信元: " . $pm->From . "\n";
-            if ($reply_to) {
-                $debug_info .= "Reply-To: " . $reply_to . "\n";
-            }
-            $debug_info .= "文字コード: " . ($this->config('mail_charset') ?: 'iso-2022-jp') . "\n";
-            $debug_info .= "HTMLメール: " . ($this->config('admin_ishtml') ? '有効' : '無効') . "\n";
-            if ($upload_flag) {
-                $debug_info .= "添付ファイル: あり\n";
-            }
-            $body_preview = mb_substr(strip_tags($pm->Body), 0, 100);
-            $debug_info .= "本文プレビュー: " . $body_preview . "...";
-            $this->debugLog($debug_info);
-        }
-
-        $sent = $pm->Send();
-        if (!$sent) {
-            $errormsg = 'メール送信に失敗しました::' . $pm->ErrorInfo;
-            $this->setError($errormsg);
-            $vars = evo()->htmlspecialchars(
-                var_export($pm, true)
-            );
-            $this->logEvent(1, 3, $errormsg . "\n" . $vars, 'Admin Mail Error');
-            return false;
-        }
-
-        $this->debugLog('管理者宛メール送信に成功しました');
-        return true;
-    }
-
-    private function sendAutoReply()
-    {
-        // 自動返信
-        $reply_to = $this->getAutoReplyAddress();
-        if (!$this->config('auto_reply') || !$reply_to) {
-            $this->debugLog('自動返信メールは無効、またはReply-Toアドレスが取得できませんでした');
-            return true;
-        }
-
-        $this->debugLog('自動返信メール送信処理を開始');
-
-        evo()->loadExtension('MODxMailer');
-        $pm = evo()->mail;
-        $pm->AddAddress($reply_to);
-        $pm->Subject = $this->config('reply_subject')
-            ? evo()->parseText($this->config('reply_subject'), $this->form)
-            : '自動返信メール';
-        $admin_addresses = $this->makeAdminAddress();
-        $admin_address = $admin_addresses[0];
-        $pm->setFrom(
-            $admin_address,
-            $this->config('reply_fromname')
-        );
-        // モバイル用のテンプレート切り替え
-        $pattern = '/(docomo\.ne\.jp|ezweb\.ne\.jp|softbank\.ne\.jp)$/';
-        if ($this->config('tmpl_mail_reply_mobile') && preg_match($pattern, $reply_to)) {
-            $template_filename = $this->config('tmpl_mail_reply_mobile');
-        } else {
-            $template_filename = $this->config('tmpl_mail_reply');
-        }
-        $tmpl_u = $this->loadTemplate($template_filename);
-        if (!$tmpl_u) {
-            $this->setError('メールテンプレートの読み込みに失敗しました');
-            return false;
-        }
-
-        $pm->Sender = $pm->From;
-        $pm->Body = $this->makeBody(
-            $tmpl_u,
-            ($this->makePh($this->form, $admin_address))
-        );
-        $pm->Encoding = '7bit';
-        // 添付ファイル処理
-        if ($this->config('attach_file') && is_file($this->config('attach_file'))) {
-            if (!$this->config('attach_file_name')) {
-                $pm->AddAttachment($this->config('attach_file'));
-            } else {
-                $pm->AddAttachment(
-                    $this->config('attach_file'),
-                    mb_convert_encoding(
-                        $this->config('attach_file_name'),
-                        $this->config('mail_charset'),
-                        $this->config('charset')
-                    )
-                );
-            }
-        }
-
-        $uploaded = sessionv('_cf_uploaded');
-        $has_uploaded_attachment = false;
-        if (is_array($uploaded)) {
-            foreach ($uploaded as $attach_file) {
-                if (!is_file($attach_file['path'])) {
-                    continue;
-                }
-                $pm->AddAttachment(
-                    $attach_file['path'],
-                    mb_convert_encoding(
-                        urldecode(basename($attach_file['path'])),
-                        $this->config('mail_charset'),
-                        $this->config('charset')
-                    )
-                );
-                $has_uploaded_attachment = true;
-            }
-        }
-
-        // デバッグログ: メール送信情報
-        if ($this->config('debug_mode')) {
-            $debug_info = "自動返信メール送信を試行\n";
-            $debug_info .= "宛先: " . $reply_to . "\n";
-            $debug_info .= "件名: " . $pm->Subject . "\n";
-            $debug_info .= "送信元: " . $pm->From . "\n";
-            if ($this->config('reply_fromname')) {
-                $debug_info .= "送信者名: " . $this->config('reply_fromname') . "\n";
-            }
-            $debug_info .= "文字コード: " . ($this->config('mail_charset') ?: 'iso-2022-jp') . "\n";
-            $debug_info .= "HTMLメール: " . ($this->config('reply_ishtml') ? '有効' : '無効') . "\n";
-            $debug_info .= "テンプレート: " . $template_filename . "\n";
-            if ($this->config('attach_file') || $has_uploaded_attachment) {
-                $debug_info .= "添付ファイル: あり\n";
-            }
-            $body_preview = mb_substr(strip_tags($pm->Body), 0, 100);
-            $debug_info .= "本文プレビュー: " . $body_preview . "...";
-            $this->debugLog($debug_info);
-        }
-
-        $sent = $pm->Send();
-
-        if (!$sent) {
-            $errormsg = 'メール送信に失敗しました::' . $pm->ErrorInfo;
-            $this->setError($errormsg);
-            $vars = evo()->htmlspecialchars(
-                var_export($pm, true)
-            );
-            $this->logEvent(1, 3, $errormsg . "\n" . $vars, 'AutoReply Error');
-            return false;
-        }
-
-        $this->debugLog('自動返信メール送信に成功しました');
-        return true;
-    }
-
     public function cleanUploadedFiles()
     {
         $uploaded = sessionv('_cf_uploaded');
@@ -867,75 +637,6 @@ class Class_cfFormMailer
             unlink($attach_file['path']);
         }
         unset($_SESSION['_cf_uploaded']);
-    }
-
-    private function makeBody($tpl, $ph)
-    {
-        $body = $this->clearPlaceHolder(
-            $this->replacePlaceHolder(
-                str_replace(
-                    array("\r\n", "\r"),
-                    "\n",
-                    $tpl
-                ),
-                $ph,
-                $this->config('allow_html') ? '<br />' : "\n"
-            )
-        );
-
-        $mail_charset = $this->config('mail_charset');
-        $charset = $this->config('charset');
-
-        // mail_charsetが空の場合はエンコード変換をスキップ
-        if (empty($mail_charset) || empty($charset)) {
-            return $body;
-        }
-
-        return mb_convert_encoding($body, $mail_charset, $charset);
-    }
-    private function makePh($form, $adminmail)
-    {
-        $remote_addr = serverv('REMOTE_ADDR', '');
-        $user_agent = serverv('HTTP_USER_AGENT', '');
-
-        $additional = array(
-            'senddate'    => date('Y-m-d H:i:s'),
-            'adminmail'   => $adminmail,
-            'sender_ip'   => $remote_addr,
-            'sender_host' => $remote_addr ? gethostbyaddr($remote_addr) : '',
-            'sender_ua'   => $this->encodeHTML($user_agent),
-            'reply_to'    => $this->getAutoReplyAddress(),
-        );
-        if (!$this->config('reply_ishtml')) {
-            return $form + $additional;
-        }
-        if (!$this->config('allow_html')) {
-            return $this->encodeHTML($form, 'true') + $additional;
-        }
-        return $this->nl2br_array($form) + $additional;
-    }
-
-    private function makeAdminAddress()
-    {
-        if (!$this->config('dynamic_send_to_field')) {
-            $mails = explode(',', $this->config('admin_mail'));
-        } else {
-            $dynamic_send_to = sessionv('dynamic_send_to');
-            $field_value = $this->form[$this->config('dynamic_send_to_field')] ?? '';
-            if (empty($dynamic_send_to) || !$field_value || !isset($dynamic_send_to[$field_value])) {
-                $mails = explode(',', $this->config('admin_mail'));
-            } else {
-                $mails = explode(',', $dynamic_send_to[$field_value]);
-            }
-        }
-        $admin_addresses = array();
-        foreach ($mails as $buf) {
-            $buf = trim($buf);
-            if ($this->_isValidEmail($buf)) {
-                $admin_addresses[] = $buf;
-            }
-        }
-        return $admin_addresses;
     }
 
     /**
@@ -1240,169 +941,6 @@ class Class_cfFormMailer
             return array_map(array($this, 'nl2br_array'), $text);
         }
         return nl2br($text);
-    }
-
-    /**
-     * テンプレートチャンクの読み込み
-     *
-     * @access private
-     * @param  string $tpl_name チャンク名・またはリソースID
-     * @return string 読み込んだデータ
-     */
-    private function loadTemplate($tpl_name)
-    {
-        $tpl_name = trim($tpl_name);
-        if (preg_match('/^@FILE:.+/', $tpl_name)) {
-            $list = array(
-                CFM_PATH . trim(substr($tpl_name, 6)),
-                MODX_BASE_PATH . trim(substr($tpl_name, 6))
-            );
-            foreach ($list as $path) {
-                if (!is_file($path)) {
-                    continue;
-                }
-                return $this->parseDocumentSource(
-                    file_get_contents($path)
-                );
-            }
-        } elseif (preg_match('/^[1-9][0-9]*$/', $tpl_name)) {
-            $doc = $this->getDocument($tpl_name);
-            if (isset($doc['content']) && $doc['content']) {
-                return $this->parseDocumentSource($doc['content']);
-            }
-        } elseif ($content = evo()->getChunk($tpl_name)) {
-            return $this->parseDocumentSource($content);
-        }
-
-        $error = 'tpl read error';
-        if ($tpl_name) {
-            $error .= sprintf(' (%s)', $tpl_name);
-        }
-        $this->setError($error);
-        return false;
-    }
-
-    private function parseDocumentSource($content)
-    {
-        if (strpos($content, '[!') !== false) {
-            $content = str_replace(array('[!', '!]'), array('[[', ']]'), $content);
-        }
-        return evo()->parseDocumentSource($content);
-    }
-
-    private function getDocument($docid)
-    {
-        $rs = db()->select(
-            '*',
-            '[+prefix+]site_content',
-            'id=' . $docid,
-            '',
-            1
-        );
-        $doc = db()->getRow($rs);
-        if (!$doc) {
-            return false;
-        }
-        return $doc;
-    }
-    /**
-     * プレースホルダを置換
-     *
-     * @access private
-     * @param  string $text HTMLテキスト
-     * @param  array  $params 置換するデータ
-     *                         (プレースホルダ名) => (値)
-     * @param  string $join 値が配列の場合に連結に使用する文字列
-     * @return string プレースホルダが置換された文字列
-     */
-    private function replacePlaceHolder($text, $params, $join = '<br />')
-    {
-        if (!is_array($params) || !$text) {
-            return false;
-        }
-
-        preg_match_all("/\[\+([^+|]+)(\|(.*?)(\((.+?)\))?)?\+]/is", $text, $match, PREG_SET_ORDER);
-        if (!count($match)) {
-            return $text;
-        }
-
-        $toFilter = true;
-        //旧バージョン用
-        if (isset(evo()->config['output_filter']) && evo()->config['output_filter'] == 0) {
-            $toFilter = false;
-        }
-
-        if ($toFilter) evo()->loadExtension('PHx') or die('Could not load PHx class.');
-
-        // 基本プレースホルダ
-        $replaceKeys = array_keys($params);
-        foreach ($match as $m) {
-            if ($toFilter && strpos($m[1], ':') !== false) {
-                $parts = explode(':', $m[1], 2);
-                $m[1] = $parts[0];
-                $modifiers = $parts[1] ?? false;
-            } else {
-                $modifiers = false;
-            }
-
-            if (!in_array($m[1], $replaceKeys)) {
-                continue;
-            }
-
-            $val = $params[$m[1]] ?? '';
-            if ($toFilter && $modifiers !== false) {
-                if ($val === '&nbsp;') {
-                    $val = '';
-                }
-                $val = evo()->filter->phxFilter($m[1], $val, $modifiers);
-                if ($val === '') {
-                    $val = '&nbsp;';
-                }
-            }
-            // テキストフィルターの処理
-            $fType = $m[3] ?? '';
-            if (empty($fType)) {
-                $text = str_replace(
-                    $m[0],
-                    is_array($val) ? implode($join, $val) : $val,
-                    $text
-                );
-                continue;
-            }
-            if (is_callable(array($this, '_f_' . $fType))) {
-                $funcName = '_f_' . $fType;
-                $fParam = $m[5] ?? '';
-                $text = str_replace(
-                    $m[0],
-                    $this->$funcName($val, $fParam),
-                    $text
-                );
-                continue;
-            }
-            if (is_callable('_filter_' . $fType)) {
-                $funcName = '_filter_' . $fType;
-                $fParam = $m[5] ?? '';
-                $text = str_replace(
-                    $m[0],
-                    $funcName($val, $fParam),
-                    $text
-                );
-                continue;
-            }
-            $text = str_replace($m[0], '', $text); // フィルター無し
-        }
-        return $text;
-    }
-
-    /**
-     * 全てのプレースホルダを削除
-     *
-     * @param string $text 対象となるHTML
-     * @return string [+variable_name+]削除後のHTML
-     */
-    private function clearPlaceHolder($text)
-    {
-        return preg_replace("@\[\+.+?\+]@uism", '', $text);
     }
 
     /**
